Fix handling of escaped metacharacters.
Due to the `__index` function on the metatable of grammar table there
was not way for `if grammar[ ESC .. c ]` to not return a value. This
caused all escaped value to generate a literal with escape character
in the literal. Eg:

    = re.compile("/-")
    _>ROOT
    |_>/-

This fix removes the metatable and instead checks for matches in the
two locations that require it.